live-preview: Have a simple mode for bools
The property editor built into the live-preview now comes with the
concept of a "simple/complex mode". There is a "K" (for complex, obviously;-)
or a "S" (for simple, just as obvious) beween the property name and property
value. you can click on the letter to switch from simple to complex mode,
but you can only switch back to simple mode if there are no complex expressions
in the line edit.

This implements a "simple" mode for bools.

# Todo

 * [ ] Find icons for simple/complex mode:-)
 * [ ] Figure out how to decide whether other types are "simple"

   To do so I will need some string functions in Slint, to e.g. detect that
   a string starts with and ends with a `"` and contains no unescaped `"`
   anywhere else to detect "simple strings". Same for colors, length, ...

 * [ ] Define more UIs for simple types